* Better support for reading remote server blueiris clip paths (it will try to detect the correct SHARE path or default to C$, D$, etc if a mapped drive is not found already)
* Initial DOODS support.   Install one of the docker images here (Be aware it is picky if you use a port other than 8080 - I tried 8084 first because 8080 was in use and the docker version still tried to use 8080 even though it was already in use)
https://github.com/snowzach/doods

Make sure its working by going to http://localhost:8080/version in a web browser first.

Then in AITTOOL, specify the ENTIRE URL for it to know it is a doods link:   http://localhost:8080/detect

+++ src/UI/ClsDoodsDetection.cs
@@ -5,7 +5,7 @@ namespace AITool
     public class ClsDoodsDetection
     {
         [JsonProperty("top")]
-        public int Top { get; set; } = 0;
+        public double Top { get; set; } = 0;
 
         [JsonProperty("left")]
         public double Left { get; set; } = 0;
